implement board component
Utilizing a double for loop I can represent the boards rows and the overall board. In the first loop I create a row array, and push into it 8 times a square. At the end of the first loop I push a div with the rows. The overall board is made from this board array. For the pieces they are given the type from the board representation 2d array.