Get pieces to move
Have function for finding out the legal moves of a piece. If the second click by a user is a coord within the legalmoves state then the move is legal. This is working so far for pawn moves forward. Still a lot of logic to make the pawns actually have accurate reports and for other pieces to work but the logic seems sound enough.

Also fixed it so that pieces can be placed back down on the same square and that turns actually work.

At the moment only pawns can be placed back down because of the way that I have set up the functions. This will be remedied as I write more logic for pieces.