Make king piece move
Added a function to check legal king moves. Generated neighbors are all the squares surrounding the intial coord square. I reduce thorugh that adding to the accum the new coordnates. I store this in tempmoves and add it to the legal moves.

Will need to account for checks / firstmove castling / moving into check / moving into the same pieces but this will work for now.