Make Knights be able to move
The same general principle as the king as been used. The knight has a set set of squares it can move to. Made the formula and reduced the coord with the array formula. Set the legal moves and viola.

Will need to account for taking pieces but jumping pieces should already be accomplished.