Make Rook piece move
At the moment I have 4 loops happening to generate the legal moves.

From the starting coordinate I want to find all for left right up and down. The reason being that I want to account for pieces being in the way in the future. The rook can potentially still move right even if the left has a piece in the way. I don't want my loop to break just because a piece is in the way of the loop. I recognize the repetition and would like to fix later.