Make Bishop move and add active squares
Pass prop down to square isActive, if the square is active utizling a before pseudoclass add a bullet. I set a 2d array of true null values which I change utilziing the legalmoves array. I change to true if the tempmoves coord.

Bishop moves similar to rook but with some differences around the logic. I have to change x and y on all moves, in any of the directions. Same idea.

Will be working with queen next, will be the same as a rook and queen