Make pawn move initial vs later moves
Passing down the moved prop to the onClick function I can check if a pawn has moved yet. This is so that a pawn can move two on its first turn and only one on any subsequent turns after. I also filtered out the current coord from places the piece can go as. May want to change the color or add a ghost piece to that square.