Rather large commit that prevents pieces from moving to illegal squares
The general premise is that I am filtering out any squares that should not have made it in. If I run into a piece of the same color, stop the loop. If I run into a piece of the opposite color stop the loop. It is a different method for all the pieces, some just required a straight up filter others required working in the loop to stop it directly.

I have also included pawn logic to make the pawn move forward and be able to take as well. Game should be more or less complete other than:

1. Checks
2. Castling
3 En passant
4 Checkmate