Fix bug with boundary pawns
When a pawn was on the boundary a user was unable to take pieces to its side. This was due to the way I did the if statement. I added a bit of logic to create a temp arrayt that I did a check for pieces of hte other color in the two neighboring squares