Make it so that you can click off pieces and fix pawn bug
On the first click if the move is illegal just restart the states to what they were before clicking. This adds the piece back down. There was a pawn bug where the loop was returing if the pawn found another pawn in front of it. Just switched up the ordering so the pawn finds the ones next to it before the one in front. For some reason I am finding an empty array : will need to check later. This has been filtered out and seems to be working now.